requirements.txt
python-digitalocean>=1.17
paramiko>=3.0
pytest>=7.0
pytest-xdist>=3.0
